Improve significantly the elastic results
It uses phonetic and autocomplete analyzers. The phonetic highlight is
preffered, excepted if the autocomplete highlighter has more tags.

Phonetic is used with edge_ngram tokenizer where autocomplete is used
with ngram. Then the phonetic result should be very good most of times.

Note that only the french stemmer is used with the phonetic analyzer.